requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.58.2
django-braces==1.15.0
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
django-tui==23.6
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
jwcrypto==1.5.0
linkify-it-py==2.0.2
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
ndg-httpsclient==0.5.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.5.0
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
textual==0.42.0
trogon==0.5.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.1.0
wrapt==1.16.0
zipp==3.17.0
